Database.Sqlite.Internal

Description

Utterly unsafe internals of the Database.Sqlite module. Useful for people who want access to the SQLite database pointer to manually call SQLite API functions via the FFI.

Types and functions in this module are *NOT* covered by the PVP and may change breakingly in any future version of the package.

Documentation

data Connection

SQLite connection type, consist of an IORef tracking whether the connection has been closed and the raw SQLite C API pointer, wrapped in a 'Connection'' newtype.

Since: 2.10.2

Constructors

Connection !(IORef Bool) Connection'

newtype Connection'

Newtype wrapping SQLite C API pointer for a database connection.

Since: 2.10.2

Constructors

Connection' (Ptr ())

newtype Statement

Newtype wrapping SQLite C API pointer for a prepared statement.

Since: 2.10.2

Constructors

Statement (Ptr ())